requirements.txt
cleverhans==3.0.1
matplotlib==3.1.3
numba==0.53.1
numpy==1.19.1
pandas==1.0.5
Pillow==6.2.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.11.1
torch==1.7.0
torchsummary==1.5.1
torchvision==0.8.1
tqdm==4.40.2
Keras==2.3.1
tensorboard==2.0.2
